fix: Profile stories ( add Provider)
